+++ tests/main.lisp
@@ -62,6 +62,8 @@
   (is equal '(2 4) (t:transduce (t:filter #'evenp) #'t:cons '(1 2 3 4 5)))
   (is equal '(2 5 8) (t:transduce (t:filter-map #'cl:first) #'t:cons '(() (2 3) () (5 6) () (8 9))))
   (is equal '(1 2 3 "abc") (t:transduce #'t:unique #'t:cons '(1 2 1 3 2 1 2 "abc")))
+  (is equal '((:a . 1) (:b . 2) (:d . 3))
+      (t:transduce (t:unique-by #'cdr) #'t:cons '((:a . 1) (:b . 2) (:c . 1) (:d . 3))))
   (is equal '(1 2 3 4 3)
       (t:transduce #'t:dedup #'t:cons '(1 1 1 2 2 2 3 3 3 4 3 3))))